turn generators into a submodule
